+++ examples/utils-test-tr.cpp
@@ -1,9 +1,5 @@
-#include <utils/tr.hpp>
+#include <utils/sys.hpp>
 
 int main(int argc,char* argv[])
 {
-    using namespace utils::tr;
-    Compiler comp = Compiler::getCompiler();
-
-    return 0;
 }