+++ examples/utils-test.cpp
@@ -181,6 +181,23 @@ void test_sys()
     {
         using namespace utils::sys;
         std::cout<<"whereis: "<<whereis("g++")<<std::endl;
+
+        Compiler comp = Compiler::getCompiler();
+        Library libf = comp.input("f.cpp")
+            .output("f")
+            .flags("-o3","-Wall")
+            .get();
+
+        if(libf.load())
+        {
+            if(libf.load_f<int,int>("print"))
+                libf["print"]->call<int>(21);
+            //cast test
+            int(*cast)(int) =  *reinterpret_cast<utils::func::Func<int,int>*>(libf["print"]);
+            cast(22);
+            
+            libf.close();
+        }
     }
 
     {